Add a way to test many BrowserStack browsers with a short command, add more browsers
